requirements.txt
opencv-python>=4.5.0
google-cloud-storage>=2.0.0
google-auth>=2.0.0
